разбивате редюсер, выделает селекторы
