+++ package.json
@@ -33,6 +33,7 @@
   },
   "dependencies": {
     "axios": "0.19.2",
+    "axios-mock-adapter": "1.18.2",
     "moment": "2.27.0",
     "prop-types": "15.7.2",
     "react": "16.13.1",
@@ -41,6 +42,7 @@
     "react-router-dom": "5.2.0",
     "redux": "4.0.5",
     "redux-mock-store": "1.5.4",
-    "redux-thunk": "2.3.0"
+    "redux-thunk": "2.3.0",
+    "reselect": "4.0.0"
   }
 }
