+++ src/hoc/with-active-tab/with-active-tab.jsx
@@ -1,4 +1,5 @@
 import React, {PureComponent} from 'react';
+import PropTypes from 'prop-types';
 import {CustomPropTypes} from "../../types";
 
 import {MovieNavList} from "../../const";
@@ -73,7 +74,10 @@ const withActiveTab = (Component) => {
   }
 
   WithActiveTab.propTypes = {
-    film: CustomPropTypes.FILM,
+    film: PropTypes.oneOfType([
+      CustomPropTypes.FILM,
+      PropTypes.bool,
+    ]),
   };
 
   return WithActiveTab;
