+++ src/reducer/app/selectors.js
@@ -1,0 +1,3 @@
+import NameSpace from '../name-space.js';
+
+export const getCurrentPage = (state) => state[NameSpace.APP].currentPage;
