+++ src/reducer/name-space.js
@@ -1,0 +1,7 @@
+const NameSpace = {
+  DATA: `DATA`,
+  SHOW: `SHOW`,
+  APP: `APP`,
+};
+
+export default NameSpace;
