+++ src/reducer/utils.js
@@ -1,0 +1,1 @@
+export const extend = (a, b) => Object.assign({}, a, b);
