+++ src/types/index.js
@@ -3,23 +3,21 @@ import PropTypes from 'prop-types';
 export const CustomPropTypes = {
   FILM: PropTypes.shape({
     id: PropTypes.number.isRequired,
-    bg: PropTypes.string.isRequired,
-    poster: PropTypes.string.isRequired,
-    src: PropTypes.string.isRequired,
     title: PropTypes.string.isRequired,
-    genre: PropTypes.string.isRequired,
-    year: PropTypes.number.isRequired,
+    poster: PropTypes.string.isRequired,
+    previewImg: PropTypes.string.isRequired,
+    bg: PropTypes.string.isRequired,
+    bgc: PropTypes.string.isRequired,
+    src: PropTypes.string.isRequired,
+    previewSrc: PropTypes.string.isRequired,
     description: PropTypes.string.isRequired,
     rating: PropTypes.number.isRequired,
     votes: PropTypes.number.isRequired,
     director: PropTypes.string.isRequired,
     starring: PropTypes.arrayOf(PropTypes.string.isRequired).isRequired,
-  }).isRequired,
-  REVIEW: PropTypes.shape({
-    id: PropTypes.string.isRequired,
-    author: PropTypes.string.isRequired,
-    date: PropTypes.string.isRequired,
-    rating: PropTypes.string.isRequired,
-    text: PropTypes.string.isRequired,
+    time: PropTypes.number.isRequired,
+    genre: PropTypes.string.isRequired,
+    year: PropTypes.number.isRequired,
+    isFavorite: PropTypes.bool.isRequired,
   }).isRequired,
 };
